requirements.txt
Flask-Cors==5.0.0
GitPython==3.1.44
Markdown==3.7
MarkupSafe==3.0.2
SQLAlchemy==2.0.37
absl-py==2.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
astunparse==1.6.3
attrs==24.3.0
blinker==1.9.0
cachetools==5.5.1
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
dateparser==1.2.0
distro==1.9.0
flask==3.1.0
flatbuffers==25.1.21
frozenlist==1.5.0
gast==0.6.0
gitdb==4.0.12
google-pasta==0.2.0
greenlet==3.1.1
groq==0.15.0
grpcio==1.69.0
h11==0.14.0
h5py==3.12.1
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
jinja2==3.1.5
joblib==1.4.2
jsonschema-specifications==2024.10.1
jsonschema==4.23.0
keras==3.8.0
libclang==18.1.1
markdown-it-py==3.0.0
mdurl==0.1.2
ml-dtypes==0.4.1
multidict==6.1.0
namex==0.0.8
narwhals==1.23.0
newsapi-python==0.2.7
nltk==3.9.1
numpy==2.0.2
opt-einsum==3.4.0
optree==0.14.0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pip==24.3.1
propcache==0.2.1
protobuf==5.29.3
pyarrow==19.0.0
pycryptodome==3.21.0
pydantic-core==2.27.2
pydantic==2.10.5
pydeck==0.9.1
pygments==2.19.1
python-binance==1.0.27
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
referencing==0.36.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
scikit-learn==1.6.1
scipy==1.15.1
setuptools==65.5.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
streamlit==1.41.1
ta==0.11.0
tenacity==9.0.0
tensorboard-data-server==0.7.2
tensorboard==2.18.0
tensorflow-intel==2.18.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow==2.18.0
termcolor==2.5.0
textblob==0.19.0
threadpoolctl==3.5.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
typing-extensions==4.12.2
tzdata==2025.1
tzlocal==5.2
urllib3==2.3.0
watchdog==6.0.0
websockets==14.2
werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
yarl==1.18.3